Only create service injector if dependencies are required
